compiler.h: add redundant cast to quiet Watcom warning
Add a redundant cast in watcom_switch_hack() to quiet a Watcom
warning, and remove open-coded implementation of the Watcom switch
hack.

+++ include/compiler.h
@@ -259,7 +259,7 @@ size_t strnlen(const char *s, size_t maxlen);
 
 static inline unsigned int watcom_switch_hack(uint64_t x)
 {
-    if (x > UINT_MAX)
+    if (x > (uint64_t)UINT_MAX)
         return BOGUS_CASE;
     else
         return (unsigned int)x;
